bgp: refactor message.rs into multiple files
